Re-organise the project structure
-removed unused / old files
- added folder structure
