+++ src/components/FindFriends.js
