+++ src/components/Stats.js
