+++ src/index.js
@@ -1,13 +1,14 @@
 import React, { useRef } from "react";
 import { createRoot } from "react-dom";
 import "./index.css";
-import Main from "./Main";
+import Main from "./components/Main";
 import firebase from "firebase";
 import "primereact/resources/themes/vela-orange/theme.css";
 import "primereact/resources/primereact.min.css";
 import "primeicons/primeicons.css";
 import { Menu } from "primereact/menu";
 import { Button } from "primereact/button";
+import * as serviceWorker from "./serviceWorker";
 
 import {
   FirebaseAppProvider,
@@ -134,3 +135,8 @@ function App() {
 // Enable Concurrent Mode
 // https://reactjs.org/docs/concurrent-mode-adoption.html#enabling-concurrent-mode
 createRoot(document.getElementById("root")).render(<App />);
+
+// If you want your app to work offline and load faster, you can change
+// unregister() to register() below. Note this comes with some pitfalls.
+// Learn more about service workers: https://bit.ly/CRA-PWA
+serviceWorker.unregister();
