+++ src/tests/App.test.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { render } from "@testing-library/react";
-import App from "./Main";
+import App from "../components/Main";
 
 test("renders learn react link", () => {
   const { getByText } = render(<App />);
